Renamed directory of RepoTable component to RepoTable.
