+++ frontend/src/Components/DatasetMetadataTable/DatasetMetadataTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import DatasetMetadataRepo from '../../Repos/DatasetMetadataRepo';
-import RepoTable from '../Shared/Table/RepoTable';
+import RepoTable from '../Shared/RepoTable/RepoTable';
 
 
 const repo = new DatasetMetadataRepo();
